Update 3.0.0 - Starboard
Fully working (at least from what i've seen during the testing) starboard that supports attached images and also with a customizable minium amount

Fixed the messageUpdate event so now it only emits a new messageCreate event when it is the last message of the user

+++ util/helpers/StarboardManager.js
@@ -20,8 +20,10 @@ class StarboardManager {
         });
         this.client.on("messageReactionUpdate", async(message, emoji, userID) => {
             message = message.channel.messages.get(message.id) || await this.client.getMessage(message.channel.id, message.id);
+            message.channel.messages.set(message.id, message);
             const guildEntry = this.client.guildData.get(message.guild.id);
             if (!guildEntry || !guildEntry.starboard.channel) return;
+            let image = message.attachments[0] ? (new RegExp(/.jpg|.jpeg|.png|.gif/gim).test(message.attachments[0].url) ? message.attachments[0].url : false) : false;
             if (!messages.get(message.id)) {
                 messages.set(message.id, {
                     id: message.id,
@@ -29,25 +31,21 @@ class StarboardManager {
                     channel: message.channel,
                     author: message.author,
                     timeoutLaunched: false,
-                    stars: message.reactions['⭐'] || message.reactions['🌟'] ?
-                        ((message.reactions['⭐'] ? message.reactions['⭐'].count : 0) + (message.reactions['🌟'] ? message.reactions['🌟'].count : 0)) :
-                        (message.reactions.filter ? message.reactions.filter(r => ['⭐', '🌟'].includes(r.emoji.name)).size : 0)
+                    image: image,
+                    stars: undefined
                 });
-            } else {
-                let inQueueMessage = messages.get(message.id);
-                inQueueMessage.stars = message.reactions['⭐'] || message.reactions['🌟'] ?
-                    ((message.reactions['⭐'] ? message.reactions['⭐'].count : 0) + (message.reactions['🌟'] ? message.reactions['🌟'].count : 0)) :
-                    (message.reactions.filter ? message.reactions.filter(r => ['⭐', '🌟'].includes(r.emoji.name)).size : 0)
-                messages.set(message.id, inQueueMessage);
             }
+            //If the timeout to update is not already launched, launch it
             if (!messages.get(message.id).timeoutLaunched) {
                 setTimeout(async() => {
-                    let inQueueMessage = messages.get(message.id);
-                    inQueueMessage.timeoutLaunched = false;
-                    messages.set(message.id, inQueueMessage);
+                    messages.get(message.id).timeoutLaunched = false;
+                    messages.get(message.id).stars = await this.getStarsCount(message);
                     let starredMessage = messages.get(message.id);
                     let storedMessage = guildEntry.starboard.messages.find(m => m.id === message.id);
-                    if (!storedMessage && starredMessage.stars) {
+                    let starboardMessageExists = storedMessage ? await this.client.getMessage(guildEntry.starboard.channel, storedMessage.starboardMessage).catch(err => false) : false;
+                    //If the message is not in the starboard yet and reach the minimum requirements, add it to the starboard
+                    if ((!storedMessage || !starboardMessageExists) && starredMessage.stars >= guildEntry.starboard.minimum) {
+                        if (storedMessage && !starboardMessageExists) guildEntry.starboard.messages.splice(guildEntry.starboard.messages.findIndex(m => m.id === message.id), 1);
                         try {
                             let starboardMessage = await this.client.createMessage(guildEntry.starboard.channel, {
                                 embed: {
@@ -60,7 +58,10 @@ class StarboardManager {
                                     footer: {
                                         text: `In #${message.channel.name}`
                                     },
-                                    timestamp: new Date().toISOString()
+                                    image: image ? {
+                                        url: image
+                                    } : undefined,
+                                    timestamp: new Date(message.timestamp).toISOString()
                                 }
                             });
                             guildEntry.starboard.messages.push({
@@ -70,11 +71,14 @@ class StarboardManager {
                                 content: starredMessage.content,
                                 author: starredMessage.author.id,
                                 stars: starredMessage.stars,
-                                timestamp: new Date().toISOString()
+                                image: image,
+                                timestamp: new Date(message.timestamp).toISOString()
                             });
                             this.client.guildData.set(message.guild.id, guildEntry);
                         } catch (err) {}
-                    } else if (storedMessage && starredMessage.stars) {
+                    }
+                    //If the message is already in the starboard and still reach the minimum requirements, update it
+                    else if (storedMessage && starredMessage.stars >= guildEntry.starboard.minimum && starboardMessageExists) {
                         if (storedMessage.starboardChannel !== guildEntry.starboard.channel) return;
                         try {
                             this.client.editMessage(storedMessage.starboardChannel, storedMessage.starboardMessage, {
@@ -88,6 +92,9 @@ class StarboardManager {
                                     footer: {
                                         text: `In #${message.channel.name}`
                                     },
+                                    image: image ? {
+                                        url: image
+                                    } : undefined,
                                     timestamp: storedMessage.timestamp
                                 }
                             });
@@ -98,11 +105,14 @@ class StarboardManager {
                                 content: storedMessage.content,
                                 author: starredMessage.author.id,
                                 stars: starredMessage.stars,
+                                image: image,
                                 timestamp: storedMessage.timestamp
                             }
                             this.client.guildData.set(message.guild.id, guildEntry);
                         } catch (err) {}
-                    } else if (storedMessage && !starredMessage.stars) {
+                    }
+                    //If the message is already in the starboard, but now under the requirements, delete it
+                    else if (storedMessage && (starredMessage.stars < guildEntry.starboard.minimum) && starboardMessageExists) {
                         try {
                             this.client.deleteMessage(storedMessage.starboardChannel, storedMessage.starboardMessage);
                             guildEntry.starboard.messages.splice(guildEntry.starboard.messages.findIndex(m => m.id === message.id), 1);
@@ -114,6 +124,20 @@ class StarboardManager {
             let inQueueMessage = messages.get(message.id);
             inQueueMessage.timeoutLaunched = true;
             messages.set(message.id, inQueueMessage);
+        });
+    }
+
+    getStarsCount(message) {
+        return new Promise(async(resolve, reject) => {
+            let starOne = await message.getReaction('⭐');
+            let starTwo = await message.getReaction('🌟');
+            let uniqueUsers = [];
+            let merged = starOne.concat(starTwo).map(u => {
+                if (!uniqueUsers.find(user => user.id === u.id)) {
+                    uniqueUsers.push(u);
+                }
+            });
+            resolve(uniqueUsers.length);
         });
     }
 }
